refactor: modularize cli and shared utilities
