+++ src/config/validation.ts
@@ -1,0 +1,30 @@
+export const SERVER_ID_PATTERN = /^[a-zA-Z0-9_-]{1,64}$/;
+export const HEADER_NAME_PATTERN = /^[!#$%&'*+\-.^_`|~0-9A-Za-z]+$/;
+export const FORBIDDEN_HEADERS = new Set([
+  "accept",
+  "authorization",
+  "connection",
+  "content-length",
+  "content-type",
+  "host",
+  "keep-alive",
+  "mcp-protocol-version",
+  "mcp-session-id",
+  "proxy-authenticate",
+  "proxy-authorization",
+  "te",
+  "trailer",
+  "transfer-encoding",
+  "upgrade",
+]);
+
+export function isAllowedRemoteUrl(value: string): boolean {
+  const url = new URL(value);
+  if (url.protocol === "https:") {
+    return true;
+  }
+  if (url.protocol !== "http:") {
+    return false;
+  }
+  return ["localhost", "127.0.0.1", "[::1]", "::1"].includes(url.hostname);
+}
